+++ src/VirtoCommerce.CartModule.Core/Services/IDeleteObsoleteCartsHandler.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+
+namespace VirtoCommerce.CartModule.Core.Services
+{
+    public interface IDeleteObsoleteCartsHandler
+    {
+        Task DeleteObsoleteCarts();
+    }
+}
